Add KoPsiDeclaration class and tests for this
