+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoBaseDeclaration.kt
@@ -8,70 +8,9 @@ import java.io.File
 /**
  * Base declaration
  */
-open class KoBaseDeclaration(private val ktElement: KtElement) {
-    /**
-     * File path of the declaration
-     */
-    val filePath by lazy {
-        ktElement
-            .containingKtFile
-            .virtualFilePath
-            .replace("//", "/")
-    }
-
-    val projectFilePath by lazy {
-        val mainPath = File("")
-            .absoluteFile
-            .path
-            .substringBeforeLast('/')
-
-        filePath.removePrefix(mainPath)
-    }
-
+open class KoBaseDeclaration(private val ktElement: KtElement): KoPsiDeclaration(ktElement) {
     /**
      * KoFile containing the declaration
      */
     val containingFile by lazy { KoFile.getInstance(ktElement.containingKtFile) }
-
-    /**
-     * Location of the declaration containing the file path, line and column
-     */
-    val location by lazy {
-        val lineAndColumn = ktElement
-            .getTextWithLocation()
-            .substringAfterLast("' at (")
-            .substringBefore(") in")
-            .split(",")
-            .toMutableList()
-            .also { it.removeIf { string -> string.isBlank() } }
-
-        val line = lineAndColumn[0]
-        val column = lineAndColumn[1]
-        "$filePath:$line:$column"
-    }
-
-    /**
-     * Text of the declaration
-     */
-    val text by lazy {
-        ktElement
-            .getTextWithLocation()
-            .substringBefore("' at (")
-            .removePrefix("'")
-            .removeSuffix("\n")
-    }
-
-    /**
-     * Text of the declaration with the location (file path, line and column).
-     */
-    val locationWithText by lazy { "Location: $location \nDeclaration:\n$text" }
-
-    fun resideInFilePath(text: String) = PackageHelper.resideInPackage(text, filePath, '/')
-
-    fun resideInProjectFilePath(text: String) = PackageHelper.resideInPackage(text, projectFilePath, '/')
-
-    /**
-     * Text of the declaration with the location (file path, line and column).
-     */
-    override fun toString() = locationWithText
 }
