+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/declaration/kopsideclaration/snippet/file-path.kttxt
