+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/declaration/kopsideclaration/snippet/location-with-single-digit.kttxt
