+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/declaration/kopsideclaration/snippet/reside-in-project-file-path.kttxt
