+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/declaration/kopsideclaration/snippet/to-string.kttxt
